PSMDB-139 Move compaction filters closer to RocksCompactionScheduler
No functional changes intended.
